waste-bot/requirements.txt
flask==2.3.3
line-bot-sdk==3.0.0
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
pillow==10.0.1
gunicorn==21.2.0